+++ ClientApp/src/app/WMS/ReceiveMaterial.component.ts
@@ -503,7 +503,7 @@ export class ReceiveMaterialComponent implements OnInit {
         return;
       }
 
-      if (totalqty == parseInt(this.matqty)) {
+      if (totalqty == parseFloat(this.matqty)) {
         this.ConfirmationService.confirm({
           message: 'Are you sure to put away material in selected stock type?',
           accept: () => {
